Moved general-purpose classes for mutation finding and application from semanticschanging to mutateapi.
