+++ generator/src/main/java/com/graphicsfuzz/generator/semanticschanging/Compound2BodyMutationFinder.java
@@ -25,6 +25,7 @@ import com.graphicsfuzz.common.ast.stmt.IfStmt;
 import com.graphicsfuzz.common.ast.stmt.Stmt;
 import com.graphicsfuzz.common.ast.stmt.WhileStmt;
 import com.graphicsfuzz.generator.mutateapi.MutationFinderBase;
+import com.graphicsfuzz.generator.mutateapi.Stmt2StmtMutation;
 import java.util.Arrays;
 
 public class Compound2BodyMutationFinder extends MutationFinderBase<Stmt2StmtMutation> {
