+++ generator/src/main/java/com/graphicsfuzz/generator/semanticschanging/Expr2BinaryMutationFinder.java
@@ -24,6 +24,8 @@ import com.graphicsfuzz.common.ast.expr.ParenExpr;
 import com.graphicsfuzz.common.ast.type.BasicType;
 import com.graphicsfuzz.common.ast.type.Type;
 import com.graphicsfuzz.common.util.IRandom;
+import com.graphicsfuzz.generator.mutateapi.Expr2ExprMutation;
+import com.graphicsfuzz.generator.mutateapi.Expr2ExprMutationFinder;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
