+++ generator/src/main/java/com/graphicsfuzz/generator/semanticschanging/Expr2LiteralMutationFinder.java
@@ -19,6 +19,8 @@ package com.graphicsfuzz.generator.semanticschanging;
 import com.graphicsfuzz.common.ast.TranslationUnit;
 import com.graphicsfuzz.common.ast.expr.Expr;
 import com.graphicsfuzz.common.util.IRandom;
+import com.graphicsfuzz.generator.mutateapi.Expr2ExprMutation;
+import com.graphicsfuzz.generator.mutateapi.Expr2ExprMutationFinder;
 
 public class Expr2LiteralMutationFinder extends Expr2ExprMutationFinder {
